src/client/requirements.txt
google-auth
requests
rich
